feat(tax): Use a customer namespace for applied taxes
