--- app/graphql/types/applied_taxes/object.rb
@@ -1,16 +1,0 @@
-# frozen_string_literal: true
-
-module Types
-  module AppliedTaxes
-    class Object < Types::BaseObject
-      graphql_name 'AppliedTax'
-
-      field :id, ID, null: false
-
-      field :customer, Types::Customers::Object, null: false
-      field :tax, Types::Taxes::Object, null: false
-
-      field :created_at, GraphQL::Types::ISO8601DateTime, null: false
-    end
-  end
-end
