--- app/models/applied_tax.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-class AppliedTax < ApplicationRecord
-  include PaperTrailTraceable
-
-  belongs_to :customer
-  belongs_to :tax
-end
