--- app/serializers/v1/applied_tax_serializer.rb
@@ -1,16 +1,0 @@
-# frozen_string_literal: true
-
-module V1
-  class AppliedTaxSerializer < ModelSerializer
-    def serialize
-      {
-        lago_id: model.id,
-        lago_customer_id: model.customer.id,
-        lago_tax_id: model.tax.id,
-        tax_code: model.tax.code,
-        external_customer_id: model.customer.external_id,
-        created_at: model.created_at.iso8601,
-      }
-    end
-  end
-end
