+++ db/migrate/20230529093955_rename_applied_taxes_to_customers_taxes.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class RenameAppliedTaxesToCustomersTaxes < ActiveRecord::Migration[7.0]
+  def change
+    rename_table :applied_taxes, :customers_taxes
+  end
+end
