+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.0].define(version: 2023_05_25_154612) do
+ActiveRecord::Schema[7.0].define(version: 2023_05_29_093955) do
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
   enable_extension "plpgsql"
@@ -85,15 +85,6 @@ ActiveRecord::Schema[7.0].define(version: 2023_05_25_154612) do
     t.integer "frequency_duration_remaining"
     t.index ["coupon_id"], name: "index_applied_coupons_on_coupon_id"
     t.index ["customer_id"], name: "index_applied_coupons_on_customer_id"
-  end
-
-  create_table "applied_taxes", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
-    t.uuid "customer_id", null: false
-    t.uuid "tax_id", null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
-    t.index ["customer_id"], name: "index_applied_taxes_on_customer_id"
-    t.index ["tax_id"], name: "index_applied_taxes_on_tax_id"
   end
 
   create_table "billable_metrics", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
@@ -282,6 +273,15 @@ ActiveRecord::Schema[7.0].define(version: 2023_05_25_154612) do
     t.index ["external_id", "organization_id"], name: "index_customers_on_external_id_and_organization_id", unique: true, where: "(deleted_at IS NULL)"
     t.index ["organization_id"], name: "index_customers_on_organization_id"
     t.check_constraint "invoice_grace_period >= 0", name: "check_customers_on_invoice_grace_period"
+  end
+
+  create_table "customers_taxes", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
+    t.uuid "customer_id", null: false
+    t.uuid "tax_id", null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["customer_id"], name: "index_customers_taxes_on_customer_id"
+    t.index ["tax_id"], name: "index_customers_taxes_on_tax_id"
   end
 
   create_table "events", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
@@ -715,8 +715,6 @@ ActiveRecord::Schema[7.0].define(version: 2023_05_25_154612) do
   add_foreign_key "add_ons", "organizations"
   add_foreign_key "applied_add_ons", "add_ons"
   add_foreign_key "applied_add_ons", "customers"
-  add_foreign_key "applied_taxes", "customers"
-  add_foreign_key "applied_taxes", "taxes"
   add_foreign_key "billable_metrics", "organizations"
   add_foreign_key "charges", "billable_metrics"
   add_foreign_key "charges", "plans"
@@ -734,6 +732,8 @@ ActiveRecord::Schema[7.0].define(version: 2023_05_25_154612) do
   add_foreign_key "credits", "invoices"
   add_foreign_key "customer_metadata", "customers"
   add_foreign_key "customers", "organizations"
+  add_foreign_key "customers_taxes", "customers"
+  add_foreign_key "customers_taxes", "taxes"
   add_foreign_key "events", "customers"
   add_foreign_key "events", "organizations"
   add_foreign_key "events", "subscriptions"
