+++ schema.graphql
@@ -93,13 +93,6 @@ type AppliedCoupon {
   terminatedAt: ISO8601DateTime!
 }
 
-type AppliedTax {
-  createdAt: ISO8601DateTime!
-  customer: Customer!
-  id: ID!
-  tax: Tax!
-}
-
 """
 Represents non-fractional signed whole numeric values. Since the value may
 exceed the size of a 32-bit integer, it's encoded as a string.
@@ -1537,18 +1530,6 @@ input CreateAppliedCouponInput {
 }
 
 """
-Autogenerated input type of CreateAppliedTax
-"""
-input CreateAppliedTaxInput {
-  """
-  A unique identifier for the client performing the mutation.
-  """
-  clientMutationId: String
-  customerId: ID!
-  taxId: ID!
-}
-
-"""
 Autogenerated input type of CreateBillableMetric
 """
 input CreateBillableMetricInput {
@@ -1602,6 +1583,18 @@ input CreateCreditNoteInput {
   items: [CreditNoteItemInput!]!
   reason: CreditNoteReasonEnum!
   refundAmountCents: BigInt
+}
+
+"""
+Autogenerated input type of CreateCustomerAppliedTax
+"""
+input CreateCustomerAppliedTaxInput {
+  """
+  A unique identifier for the client performing the mutation.
+  """
+  clientMutationId: String
+  customerId: ID!
+  taxId: ID!
 }
 
 """
@@ -2530,7 +2523,7 @@ type Customer {
   applicableTimezone: TimezoneEnum!
   appliedAddOns: [AppliedAddOn!]
   appliedCoupons: [AppliedCoupon!]
-  appliedTaxes: [AppliedTax!]
+  appliedTaxes: [CustomerAppliedTax!]
   billingConfiguration: CustomerBillingConfiguration
 
   """
@@ -2594,6 +2587,13 @@ type Customer {
   url: String
   vatRate: Float
   zipcode: String
+}
+
+type CustomerAppliedTax {
+  createdAt: ISO8601DateTime!
+  customer: Customer!
+  id: ID!
+  tax: Tax!
 }
 
 type CustomerBillingConfiguration {
@@ -2660,17 +2660,6 @@ type DestroyAddOnPayload {
 }
 
 """
-Autogenerated input type of DestroyAppliedTax
-"""
-input DestroyAppliedTaxInput {
-  """
-  A unique identifier for the client performing the mutation.
-  """
-  clientMutationId: String
-  id: ID!
-}
-
-"""
 Autogenerated input type of DestroyBillableMetric
 """
 input DestroyBillableMetricInput {
@@ -2712,6 +2701,17 @@ type DestroyCouponPayload {
   """
   clientMutationId: String
   id: ID
+}
+
+"""
+Autogenerated input type of DestroyCustomerAppliedTax
+"""
+input DestroyCustomerAppliedTaxInput {
+  """
+  A unique identifier for the client performing the mutation.
+  """
+  clientMutationId: String
+  id: ID!
 }
 
 """
@@ -3236,16 +3236,6 @@ type Mutation {
   ): AppliedCoupon
 
   """
-  Assign a tax to a customer
-  """
-  createAppliedTax(
-    """
-    Parameters for CreateAppliedTax
-    """
-    input: CreateAppliedTaxInput!
-  ): AppliedTax
-
-  """
   Creates a new Billable metric
   """
   createBillableMetric(
@@ -3284,6 +3274,16 @@ type Mutation {
     """
     input: CreateCustomerInput!
   ): Customer
+
+  """
+  Assign a tax to a customer
+  """
+  createCustomerAppliedTax(
+    """
+    Parameters for CreateCustomerAppliedTax
+    """
+    input: CreateCustomerAppliedTaxInput!
+  ): CustomerAppliedTax
 
   """
   Creates a new Customer Wallet
@@ -3376,16 +3376,6 @@ type Mutation {
   ): DestroyAddOnPayload
 
   """
-  Unassign a tax from a customer
-  """
-  destroyAppliedTax(
-    """
-    Parameters for DestroyAppliedTax
-    """
-    input: DestroyAppliedTaxInput!
-  ): AppliedTax
-
-  """
   Deletes a Billable metric
   """
   destroyBillableMetric(
@@ -3414,6 +3404,16 @@ type Mutation {
     """
     input: DestroyCustomerInput!
   ): DestroyCustomerPayload
+
+  """
+  Unassign a tax from a customer
+  """
+  destroyCustomerAppliedTax(
+    """
+    Parameters for DestroyCustomerAppliedTax
+    """
+    input: DestroyCustomerAppliedTaxInput!
+  ): CustomerAppliedTax
 
   """
   Destroy a payment provider
