+++ spec/factories/customer_applied_taxes.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 FactoryBot.define do
-  factory :applied_tax do
+  factory :customer_applied_tax, class: 'Customer::AppliedTax' do
     customer
     tax
   end
