--- spec/models/applied_tax_spec.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-
-RSpec.describe AppliedTax, type: :model do
-  subject(:applied_tax) { create(:applied_tax) }
-
-  it_behaves_like 'paper_trail traceable'
-end
